+++ substrate-archive-backend/Cargo.toml
@@ -37,10 +37,5 @@ sp-trie = { git = "https://github.com/paritytech/substrate", branch = "master" }
 sp-version = {  git = "https://github.com/paritytech/substrate", branch = "master" }
 sp-wasm-interface = {  git = "https://github.com/paritytech/substrate", branch = "master" }
 
-# Workspace
-substrate-archive-common = { path = '../substrate-archive-common' }
-
 [features]
-default = ["logging"]
-logging = ["substrate-archive-common/logging"]
 test_rocksdb = []
