+++ substrate-archive/Cargo.toml
@@ -43,6 +43,7 @@ sp-block-builder = { git = "https://github.com/paritytech/substrate", branch = "
 sp-core = { git = "https://github.com/paritytech/substrate", branch = "master" }
 sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "master" }
 sp-tracing = { git ="https://github.com/paritytech/substrate", branch = "master" }
+sp-storage = { git ="https://github.com/paritytech/substrate", branch = "master" }
 
 # Workspace
 substrate-archive-backend = { path = '../substrate-archive-backend' }
