+++ substrate-archive/src/database/queries.rs
@@ -24,9 +24,7 @@ use sqlx::PgConnection;
 
 use sp_runtime::traits::Block as BlockT;
 
-use substrate_archive_common::models::BlockModel;
-
-use crate::error::Result;
+use crate::{database::models::BlockModel, error::Result};
 
 /// Return type of queries that `SELECT version`
 struct Version {
